updated in-memory sample
