+++ AddingDataSources/InMemory/client/index.html
@@ -15,23 +15,22 @@
     <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.0/dist/jquery.min.js" ></script>
     <script src="https://cdn.quilljs.com/1.3.6/quill.min.js" ></script>    
     <script src="https://unpkg.com/dayjs@1.8.21/dayjs.min.js" ></script>    
-    <script src="https://dl.revealbi.io/reveal/libs/1.1.7/infragistics.reveal.js"></script>   
+    <script src="https://dl.revealbi.io/reveal/libs/1.3.0/infragistics.reveal.js"></script>   
 
     <script type="text/javascript">
         
         //set this to your server url
-        $.ig.RevealSdkSettings.setBaseUrl("https://localhost:7111/");
+        $.ig.RevealSdkSettings.setBaseUrl("http://localhost:5111/");
     
-        $.ig.RevealSdkSettings.ensureFontsLoadedAsync().then(() => {
-            var revealView = new $.ig.RevealView("#revealView");
-            revealView.onDataSourcesRequested = (callback) => {
+        var revealView = new $.ig.RevealView("#revealView");
+        revealView.startInEditMode = true;
+        revealView.onDataSourcesRequested = (callback) => {
 
-                var inMemoryDataSourceItem = new $.ig.RVInMemoryDataSourceItem("SalesRecords");
-                inMemoryDataSourceItem.title = "Sales Records";
+            var inMemoryDataSourceItem = new $.ig.RVInMemoryDataSourceItem("SalesRecords");
+            inMemoryDataSourceItem.title = "Sales Records";
 
-                callback(new $.ig.RevealDataSources([], [inMemoryDataSourceItem], true));
-            };
-        });
+            callback(new $.ig.RevealDataSources([], [inMemoryDataSourceItem], true));
+        };
     </script>
 </body>
