--- AddingDataSources/InMemory/server/RevealSdk.Server/Business/Product.cs
@@ -1,8 +1,0 @@
-﻿namespace RevealSdk.Server.Business
-{
-    public class Product
-    {
-        public string Name { get; set; } = string.Empty;
-        public double UnitPrice { get; set; }
-    }
-}
