--- AddingDataSources/InMemory/server/RevealSdk.Server/Business/Seller.cs
@@ -1,8 +1,0 @@
-﻿namespace RevealSdk.Server.Business
-{
-    public class Seller
-    {
-        public string Name { get; set; } = string.Empty;
-        public string City { get; set; } = string.Empty;
-    }
-}
