+++ AddingDataSources/InMemory/server/aspnet/RevealSdk.Server/Business/Product.cs
@@ -1,0 +1,8 @@
+﻿namespace RevealSdk.Server.Business
+{
+    public class Product
+    {
+        public string Name { get; set; } = string.Empty;
+        public double UnitPrice { get; set; }
+    }
+}
