+++ AddingDataSources/InMemory/server/aspnet/RevealSdk.Server/Business/Seller.cs
@@ -1,0 +1,8 @@
+﻿namespace RevealSdk.Server.Business
+{
+    public class Seller
+    {
+        public string Name { get; set; } = string.Empty;
+        public string City { get; set; } = string.Empty;
+    }
+}
